v5 ajout ifelse pour la suppression si l'utilisateur n'existe pas

+++ mcp_users.sh
@@ -21,10 +21,16 @@ fi
 ;;
 2)cat /etc/passwd
 read -p "Entrez le nom d'utilisateur à supprimer : " userName
+if [ -n "$(grep $userName /etc/passwd)" ]
+then
 sudo userdel $userName
 echo "$userName a été supprimé"
+else
+echo "L'utilisateur n'existe pas, recommencez : "
+./mcp_users.sh
+fi
 ;;
 3)exit 0;;
 *) echo "Choix invalide, recommencez : "
 ./mcp_users.sh;;
-esac
+esac
